split up utility_core.* into file_system_for_test.* and thread-factory_for_test.*.
