+++ test/test_common/file_system_for_test.h
@@ -1,0 +1,9 @@
+#include "envoy/filesystem/filesystem.h"
+
+namespace Envoy {
+
+namespace Filesystem {
+Instance& fileSystemForTest();
+} // namespace Filesystem
+
+} // namespace Envoy
