+++ test/test_common/thread_factory_for_test.cc
@@ -1,5 +1,4 @@
 #include "common/common/thread_impl.h"
-#include "common/filesystem/filesystem_impl.h"
 
 namespace Envoy {
 
@@ -16,20 +15,5 @@ ThreadFactory& threadFactoryForTest() {
 }
 
 } // namespace Thread
-
-namespace Filesystem {
-
-// TODO(sesmith177) Tests should get the Filesystem::Instance from the same location as the main
-// code
-Instance& fileSystemForTest() {
-#ifdef WIN32
-  static InstanceImplWin32* file_system = new InstanceImplWin32();
-#else
-  static InstanceImplPosix* file_system = new InstanceImplPosix();
-#endif
-  return *file_system;
-}
-
-} // namespace Filesystem
 
 } // namespace Envoy
