+++ test/test_common/thread_factory_for_test.h
@@ -1,4 +1,3 @@
-#include "envoy/filesystem/filesystem.h"
 #include "envoy/thread/thread.h"
 
 namespace Envoy {
@@ -6,9 +5,5 @@ namespace Envoy {
 namespace Thread {
 ThreadFactory& threadFactoryForTest();
 } // namespace Thread
-
-namespace Filesystem {
-Instance& fileSystemForTest();
-} // namespace Filesystem
 
 } // namespace Envoy
