+++ test/test_common/utility.h
@@ -24,8 +24,9 @@
 #include "common/stats/fake_symbol_table_impl.h"
 #include "common/stats/raw_stat_data.h"
 
+#include "test/test_common/file_system_for_test.h"
 #include "test/test_common/printers.h"
-#include "test/test_common/utility_core.h"
+#include "test/test_common/thread_factory_for_test.h"
 
 #include "absl/time/time.h"
 #include "gmock/gmock.h"
